refactor(server): extract request dispatch into server_dispatch.go
Move handleRequest, handleRequestForFunction, handleError out of
server.go into server_dispatch.go. Same package, mechanical move.
